Fix use of errgroup.Group in main function
Previously, we just used a zero-value `errgroup.Group` to wait for server shutdown. This has the problem that if startup fails, the command hangs indefinitely, because the context was never cancelled and this `Stop` was never called on the server.

Now, we wrap the context returned from `signal.NotifyContext` with a call to `errgroup.WithContext`. The returned context from the latter is cancelled _the first time_ any function passed to the error group returns with an error, which makes everything work correctly.

For this to work, both server startup and shutdown are now passed to the error group.